Cleanup of unused code
Delete config dir as this is not used at all.
Fixing part of issue #55.

Delete osp directory as this is not used now and if it will be we will
move it to different config file.

Deleting ocs-ci.yaml.template as this is part of old runner from which
we are moving to py.test and such config was moved to different place
with --ocsci-conf param.