Fix pull-secret handling
openshift-install expects the pullSecret value to be a JSON string. Mine is
stored as a multiline object, and so the existing code to read it failed. I
really think a better approach is to use the existing parsers for both JSON and
YAML here, as they are extremely well-tested and supported.

Signed-off-by: Zack Cerza <zack@redhat.com>