Implement Config object as dataclass
The existing "set toplevel module values" approach makes it impossible
to reset the project's configuration between unit tests. As unit tests
are supposed to be atomic (read: not influence anything outside their
scope), this presents a problem. With this implementation, one unit test
 need not affect any other. Integration tests can choose not to reset
the configuration if they desire that behavior.

Signed-off-by: Zack Cerza <zack@redhat.com>