check when a pod has no containerStatuses info
When a pod in storage namespace has no status.containerStatuses info in
it's yaml representation, the pod is skipped and the fact is logged as
a warning. Eg.:

```
WARNING:ocs_ci.ocs.version:pod noobaa-core-0 has no containerStatuses
```

Note that such warning is visible only when `report-vresion` tool is
executed with --loglevel option, as the logging messages are not shown
by default.

This prevents the following error:

```
Traceback (most recent call last):
  File "/home/ocsqe/projects/ocs-ci-version/venv/bin/report-version", line 11, in <module>
    load_entry_point('ocs-ci', 'console_scripts', 'report-version')()
  File "/home/ocsqe/projects/ocs-ci-version/ocs_ci/ocs/version.py", line 167, in main
    cluster_version, image_dict = get_ocs_version()
  File "/home/ocsqe/projects/ocs-ci-version/ocs_ci/ocs/version.py", line 86, in get_ocs_version
    for cs in pod['status']['containerStatuses']:
KeyError: 'containerStatuses'
```

Signed-off-by: Martin Bukatovič <mbukatov@redhat.com>