Add retry in get function for wait for resource
Fixes: #929
During attempts for waiting for resource, it can happen that the
resource is deleted beteween the quieries and then the get function fail.
This fix adding additional retry to get function which is used in
wait_for_resource function.

This makes function more robust and proof for such issues whene the
resource is deleted and recreated between the calls.

The origin behavior for get function remain the same with no extra
attempt/retry or wait.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>