Add retry to setup fio function
Fixes: #1200

This adds retry the setup function for fio.
Currently upgrade is failing on this issue and we need unblock upgrade
jobs ASAP hence adding retry as quick fix here.

For sure Vasus propsed PR to use preinstalled images is still valid but
this setup is executed all the time as well to make sure we have the
fio package.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>