change markers of test_monitoring_defaults
Replaced deployment with post_deployment marker, so that this test will
be still executed right after installation, but it won't be part of the
deployment itself.

Until the plumbing related to post_deployment marker is done, this change
doesn't ensure the post installation priority for the test.

Based on discussion with dahorak.

Signed-off-by: Martin Bukatovič <mbukatov@redhat.com>