Fix the get_az_count() function to default to 3.
When the configuration sets none of the variables
worker_availability_zones, availability_zone_count,
or platform is not equal to 'vmware', then the get_az_count()
function currently returns "3" as a default.

When there are actually less than 3 zones, this leads to
errors in the post-install deployment checks that check
for the failure domain being set to "zone" in the crush map.

The safe default is one zone.

Signed-off-by: Michael Adam <obnox@redhat.com>