custom check replacing default query validation
Fixture measure_stop_ceph_mon removes one ceph mon from the cluster, and
when OCS cluster recovers from this, a new mon pod is deployed, while
the original no longer exists.

Example of what is happening, assuming the fixture is going to remove
mon.c:

- ceph_mon_quorum_status{ceph_daemon="mon.c"} is 1 (ok)
- fixture removes mon.c
- ceph_mon_quorum_status{ceph_daemon="mon.c"} is 0 (a problem as
  expected)
- fixture fixes the cluster to run 3 mons again, mon.c no longer exists,
  but mon.d is up and running
- ceph_mon_quorum_status{ceph_daemon="mon.c"} no longer has any values
- ceph_mon_quorum_status{ceph_daemon="mon.d"} is 1 (ok)

This combined with the change done as a part of
https://github.com/red-hat-storage/ocs-ci/pull/1690 means that
a query range of ceph_mon_quorum_status in test_monitoring_shows_mon_down
will be missing values in the last 2 minutes of so, which would normally
make the query validation implemented in query_range() method fail.

To resolve this for test_monitoring_shows_mon_down:

- the validation in query_range() was disabled in this test
- custom check was added to the test, which cares about min. number of
  datapoints based on min_downtime value

Signed-off-by: Martin Bukatovič <mbukatov@redhat.com>