Fix issue 2147
Previous NamedTemporaryFile calls saved the entrire file
object and did not release the file descriptor.  Now the
calls only save the name of the file (the only data the
code used from the file object).  The NamedTemporaryFile
call is made within a with clause to insure that the
file descriptor is released.

Signed-off-by: wusui <wusui@redhat.com>