Add interface and implementation for adding nodes and storage
This is an effort to standardize on an interface that can be used
by any component to implement add node and add storage to nodes in
an existing OCP/OCS cluster. This will work in conjunction with
platform_nodes interface.

This PR introduces NodeManagement interface and also an implementation
of the same for IBM Power platform.

Signed-off-by: gitsridhar <svenkat@us.ibm.com>