Handle remote execution scenario for services
When ocs-ci is executed locally on infra or bastion server, it can
ssh into nodes and perform service stop or start or gather status.
However, when ocs-ci is setup to run from remote system, this is not
going to work. This PR proposes the ssh command is executed with jump
server and the jump server is configured as bastion_ip in the ocs-ci
configuration. In simple scenario, the bastion_ip in the ocs-ci config
will point to bastion server of the OCP cluster - from where the nodes
are accessible.

Signed-off-by: gitsridhar <svenkat@us.ibm.com>