get_openshift_client: Avoid removing unnecessarily
On some platforms (e.g. ppc64le), we deploy OCS on top of an OCP that
was deployed with a very different toolchain from ocs-ci. We don't have
quite the same level of control over the exact OCP version.
get_openshift_client's current behavior is problematic here, especially
since it only knows how to find amd64 binaries. If skip_ocp_deployment
is set, we ought to assume that the oc binary is safe to use.

Signed-off-by: Zack Cerza <zack@redhat.com>