Remove cpu and memory limits for ceph pods
cpus are platform specifici:
-  1 cpu is 1/2 of a processor core for x86
-  1 cpu is 1/8 of a processor core for power

The ocs-ci suite includes a performance test, which seems contrary to the goal
of measuring performance.  These limits would have to be scaled to make sense for
power.  But, is it desirable to have limits on a storage resource?  A limit on
storage performance is an indirect limit on workload performance.

The UI deployment of ocs does not impose cpu and memory limits.  Shouldn't ocs-ci
emulate customer configuration?   I don't think it is needed or desirable.

By default, the linux scheduler and virtual memory management provide fairness.
The use of limits makes sense when a subsystem is cpu intensive or memory intensive,
but ceph is neither.

Signed-off-by: Luke Browning <lukebrowning@us.ibm.com>