drop failure-domain.beta.kubernetes.io/zone usage
And also make sure ocs-ci code uses constants.ZONE_LABEL with the final
(non beta) label.

Zone label failure-domain.beta.kubernetes.io/zone was deprecated in
Kubernetes 1.17.1 (since OCP 4.4.0). Another reason to use this
deprecated zone label was now fixed OCS BZ 1915807. So there is no need
to keep this in the code (we don't plan to test with OCP 4.3, neither we
want to keep workaround for a bug which was fixed before GA).

Signed-off-by: Martin Bukatovič <mbukatov@redhat.com>