Update how to get build ID in reports
As internal build id will not be in CSV we need to replace the way how
we are reporting internal build ID in mail report.

As a quick solution we can get version from tag of the catalog source
registry image. Boris is working on the way to get the version to the
CSV, maybe as a label, but till it's implemented we will rely on the
tag from the registry image.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>