Fix for MCG verification deployment on external cluster
It seems that the MCG deployment is not ready at the time we check it
in the case of external deployment.

See BZ: https://bugzilla.redhat.com/show_bug.cgi?id=2014026
For more details.

To adding re-try also for KeyError when OCP doesn't return state and
phase of the object when it's in early stages of deployment we need to
do re-try to get those data on other attempt instead of fail the whole
deployment. As those object exists few moments later then we are running
verification.

Fixes: #5186

Signed-off-by: Petr Balogh <pbalogh@redhat.com>