Add RHEL8 repos
Also adding support for changing between RHEL 7 or RHEL8.
For now, the ansible plabook for adding RHEL 8 nodes is documented
to run on RHEL 7 node. So we need to keep both repos for now.

Also the function for creating RHEL pod is not parametrized to create
different version of RHEL than RHEL 7. Once we will need to move to
RHEL 8 pod, the function needs to be updated as well.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>