Longevity Testing: Stage0 and Stage1
In Longevity testing we make sure that the ODF can continue
responding to the user and admin operations reliably and consistently under
significant load for an extended period.

To achieve this the tests are categorized into below stages.

Stage-0: Creates the initial softconfiguration of all supported types
on the Longevity cluster and keep the created resources stay forever on the cluster
Stage-1: Configure and run OCP and Application workloads. These workloads
will be run continuously for a specified period of time
Stage-2: Sequential PVC, OBC, APP pod lifecycle operations for a specified period of time
Stage-3: Concurrent PVC, OBC, APP pod lifecycle	operations for a specified period of time
Stage-4: Clone, Snapshot, Expand operations for a specified period of time

This PR automates Longevity Stage-0 and Stage-1

Signed-off-by: Prasad Desala <tdesala@redhat.com>