Rename rosa deployment configs with stable inead of stage iname
the name ofcan be used to understand addon type itself
in managed service we have QE addon and stable addon in stagend
stable addon in production. So refering the name stage in filename
is not appropiate to derstand ich config to use for production.
So just renaming file for stable addon indication

Signed-off-by: suchita.gatfane <sgatfane@redhat.com>