Pass blobs as a Python List for deletion
`GCPBucket.delete_blobs` can't process the default iterator object
that is provided by `list_blobs`. So the options are either to
iterate over it and delete the blobs individually, or to pass it as a
Python List instead. Both options are similar in their time complexity,
but the second is shorter and thus easier to maintain.

Signed-off-by: Sagi Hirshfeld <shirshfe@redhat.com>