use SC with vSAN Default Storage Policy as default for ODF deployment
Description:

Since OCP deployment comes with thin-csi as default SC and it has
StoragePolicyName as "openshift-storage-policy-<clustername>". ODF
deployment uses thin-csi for deployment which inturn creaes RAID 1 disks
for osd and mon. To save storage on vSphere, create new SC which has same param
as thin-csi and use that for ODF deployment.

Signed-off-by: vavuthu <vavuthu@redhat.com>