Fix Scale ceph pod respin test while loop is not exited
Issues Observed: github issue #2201
PVC not reaching Bound state due to issues in environment like Ceph Cluster health not in good shape
POD not reaching Running state due to lack of resources for running pods in cluster

In above scenarios TC while loop is continued till it creates the required number of resources(PVCs and PODs) in the cluster, due to this TC continued to execute for more than 4 days.

In the PR will be adding a iteration counter to terminate the while loop.

Signed-off-by: Ramakrishnan Periyasamy <rperiyas@redhat.com>