Adjust pytest_collection to config arg requirement
Trying to set a different argument name than config
for the pytest_collection hook results in an error
when running pytest:

pluggy._manager.PluginValidationError

This patch changes the argument name to config,
and accesses the global ocs-ci config object
in the function via:

ocs_config = globals()["config"]

The constant DECORATORS_CHECK_IGNORED_MARKERS
is also added to the constants.py file. In master
this is called SQUAD_CHECK_IGNORED_MARKERS, but
the name was changed since squad ownership is not
verified in release branches prior to 4.14.

Finally, some indentation and formatting was
automatically enforced by Black conftest.py.

Signed-off-by: Sagi Hirshfeld <shirshfe@redhat.com>