Enhance logic for ceph_health_recovery to update Jira
This enhancement allow us to specify the specific pattern or furnction
to confirm the jira issue with the specific ceph health issue and
comment to the bug directly via our automation account.

Also this allow us to ignore well know issues in deployments and
stabilize our jobs which were able to recover ceph health on known issue
but don't have it unnoticed because of automatic comment in the bug
itself.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>