W/A for multi platform deployment issue
We have serious issue in design and how the deployment is implemented
for multicluster scenarios as only single deployer class object is
initiated and then we cannot rely on it when we iterating over multiple
clusters in underlaying deployment methods for do_deploy_rdr and so on.

Issue:
https://issues.redhat.com/browse/OCSQE-4214

I am tryin to W/A this on curcial parts to rely on config values which
should be specific for concrete cluster.

But to fix this properly will require bigger refactoring and broader
discussion with team.

Signed-off-by: Petr Balogh <pbalogh@redhat.com>